requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Recommendation Engine
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# Analytics & Visualization
streamlit==1.30.0
plotly==5.18.0
pandas==2.2.0

# Email
jinja2==3.1.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
